Factor out array casting to own module
